@@ -3,13 +3,11 @@
 package main
 
 import (
-	"fmt"
+	"errors"
 	"net"
 	"syscall"
 	"time"
 )
-
-// TODO provide implementation based on BPF
 
 type RawAddr struct {
 	HardwareAddr net.HardwareAddr
@@ -23,19 +21,19 @@ type RawConn struct {
 }
 
 func NewRawConn(bind *RawAddr) (rc *RawConn, err error) {
-	return nil, fmt.Errorf("not implemented")
+	return nil, errors.New("not implemented")
 }
 
 func (rc *RawConn) SetReadDeadline(deadline time.Time) error {
-	return fmt.Errorf("not implemented")
+	return errors.New("not implemented")
 }
 
 func (rc *RawConn) ReadFrom(data []byte) (read int, from *RawAddr, err error) {
-	return 0, nil, fmt.Errorf("not implemented")
+	return 0, nil, errors.New("not implemented")
 }
 
 func (rc *RawConn) WriteTo(from, to *RawAddr, data []byte) (written int, err error) {
-	return 0, fmt.Errorf("not implemented")
+	return 0, errors.New("not implemented")
 }
 
 func BindToDevice(handle int, name string) error {
