@@ -5,11 +5,13 @@ package main
 import (
 	"bytes"
 	"encoding/binary"
-	"fmt"
+	"errors"
 	"net"
 	"os"
 	"syscall"
 	"time"
+
+	"github.com/pyke369/golang-support/ustr"
 )
 
 type RawAddr struct {
@@ -50,19 +52,19 @@ func NewRawConn(bind *RawAddr) (rc *RawConn, err error) {
 		rc.bind.HardwareAddr, _ = net.ParseMAC("ff:ff:ff:ff:ff:ff")
 	}
 	if rc.bind.HardwareAddr == nil {
-		return nil, fmt.Errorf("invalid bind hardware address")
+		return nil, errors.New("invalid bind hardware address")
 	}
 	if rc.bind.Addr == nil {
 		rc.bind.Addr = net.IPv4bcast
 	}
 	if rc.bind.Addr == nil {
-		return nil, fmt.Errorf("invalid bind address")
+		return nil, errors.New("invalid bind address")
 	}
 	if rc.bind.Addr.To4() == nil {
 		rc.version = 6
 	}
 	if rc.bind.Port < 0 || rc.bind.Port > 65535 {
-		return nil, fmt.Errorf("invalid bind port %d", rc.bind.Port)
+		return nil, errors.New("invalid bind port " + ustr.Int(rc.bind.Port))
 	}
 	ethertype := syscall.ETH_P_IP
 	if rc.version == 4 {
@@ -70,33 +72,38 @@ func NewRawConn(bind *RawAddr) (rc *RawConn, err error) {
 		if rc.handle, err = syscall.Socket(syscall.AF_PACKET, syscall.SOCK_RAW, ethertype); err != nil {
 			return nil, err
 		}
+
 	} else {
 		ethertype = (syscall.ETH_P_IPV6 << 8) | (syscall.ETH_P_IPV6 >> 8)
 		if rc.handle, err = syscall.Socket(syscall.AF_PACKET, syscall.SOCK_RAW, ethertype); err != nil {
 			return nil, err
 		}
 	}
-	if err = syscall.SetsockoptInt(rc.handle, syscall.SOL_SOCKET, syscall.SO_REUSEADDR, 1); err != nil {
+	if err := syscall.SetsockoptInt(rc.handle, syscall.SOL_SOCKET, syscall.SO_REUSEADDR, 1); err != nil {
 		return nil, err
 	}
-	if err = syscall.SetNonblock(rc.handle, true); err != nil {
+	if err := syscall.SetNonblock(rc.handle, true); err != nil {
 		return nil, err
 	}
 	if rc.bind.Device != "" {
 		if iface, err := net.InterfaceByName(rc.bind.Device); err != nil {
-			return nil, fmt.Errorf("invalid bind device: %v", err)
+			return nil, errors.New("invalid bind device: " + err.Error())
+
 		} else if iface.HardwareAddr == nil {
-			return nil, fmt.Errorf("no hardware address for interface %s", rc.bind.Device)
+			return nil, errors.New("no hardware address for interface " + rc.bind.Device)
+
 		} else {
 			rc.Local.HardwareAddr = iface.HardwareAddr
 			if addresses, err := iface.Addrs(); err != nil {
 				return nil, err
+
 			} else {
 				for _, address := range addresses {
 					if value, ok := address.(*net.IPNet); ok {
 						if rc.version == 4 && value.IP.To4() != nil {
 							rc.Local.Addr = value.IP
 							break
+
 						} else if rc.version == 6 && value.IP.To4() == nil {
 							rc.Local.Addr = value.IP
 							break
@@ -104,7 +111,7 @@ func NewRawConn(bind *RawAddr) (rc *RawConn, err error) {
 					}
 				}
 			}
-			if err = syscall.Bind(rc.handle, &syscall.SockaddrLinklayer{Protocol: uint16(ethertype), Ifindex: iface.Index}); err != nil {
+			if err := syscall.Bind(rc.handle, &syscall.SockaddrLinklayer{Protocol: uint16(ethertype), Ifindex: iface.Index}); err != nil {
 				return nil, err
 			}
 			rc.Local.Device = rc.bind.Device
@@ -114,8 +121,8 @@ func NewRawConn(bind *RawAddr) (rc *RawConn, err error) {
 			}
 		}
 	}
-	if rc.conn = os.NewFile(uintptr(rc.handle), fmt.Sprintf("rawconn%d", rc.handle)); rc.conn == nil {
-		return nil, fmt.Errorf("raw conn creation failed")
+	if rc.conn = os.NewFile(uintptr(rc.handle), "rawconn"+ustr.Int(rc.handle)); rc.conn == nil {
+		return nil, errors.New("raw conn creation failed")
 	}
 	return
 }
@@ -135,7 +142,7 @@ func (rc *RawConn) ReadFrom(data []byte) (read int, from *RawAddr, err error) {
 		from = &RawAddr{HardwareAddr: net.HardwareAddr{}, Device: rc.Local.Device}
 		from.HardwareAddr = append(from.HardwareAddr, data[6:12]...)
 		to := RawAddr{HardwareAddr: net.HardwareAddr{}}
-		to.HardwareAddr = append(from.HardwareAddr, data[:6]...)
+		to.HardwareAddr = append(to.HardwareAddr, data[:6]...)
 		if rc.version == 4 {
 			if read < 42 || data[23] != syscall.IPPROTO_UDP {
 				continue
@@ -147,6 +154,7 @@ func (rc *RawConn) ReadFrom(data []byte) (read int, from *RawAddr, err error) {
 			to.Port = int(binary.BigEndian.Uint16(data[14+hsize+2:]))
 			copy(data, data[14+hsize+8:])
 			read -= 14 + hsize + 8
+
 		} else {
 			if read < 62 || data[20] != syscall.IPPROTO_UDP {
 				continue
@@ -173,7 +181,7 @@ func (rc *RawConn) ReadFrom(data []byte) (read int, from *RawAddr, err error) {
 
 func (rc *RawConn) WriteTo(from, to *RawAddr, data []byte) (written int, err error) {
 	if to == nil || to.Port == 0 {
-		return 0, fmt.Errorf("invalid destination port")
+		return 0, errors.New("invalid destination port")
 	}
 	if from == nil {
 		from = &RawAddr{HardwareAddr: rc.Local.HardwareAddr, Addr: rc.Local.Addr, Port: rc.bind.Port}
@@ -182,13 +190,13 @@ func (rc *RawConn) WriteTo(from, to *RawAddr, data []byte) (written int, err err
 		from.HardwareAddr = rc.Local.HardwareAddr
 	}
 	if from.HardwareAddr == nil {
-		return 0, fmt.Errorf("invalid source hardware address")
+		return 0, errors.New("invalid source hardware address")
 	}
 	if from.Port == 0 {
 		from.Port = rc.bind.Port
 	}
 	if from.Port == 0 {
-		return 0, fmt.Errorf("invalid source port")
+		return 0, errors.New("invalid source port")
 	}
 
 	payload := make([]byte, 0, 62+len(data))
@@ -202,12 +210,13 @@ func (rc *RawConn) WriteTo(from, to *RawAddr, data []byte) (written int, err err
 		if from.Addr == nil {
 			from.Addr = net.IPv4zero
 		}
+
 	} else {
 		if to.Addr == nil {
 			to.Addr = net.IPv6linklocalallrouters
 		}
 		if to.HardwareAddr == nil {
-			to.HardwareAddr, _ = net.ParseMAC(fmt.Sprintf("33:33:%02x:%02x:%02x:%02x", to.Addr[12], to.Addr[13], to.Addr[14], to.Addr[15]))
+			to.HardwareAddr, _ = net.ParseMAC(ustr.Hex([]byte{0x33, 0x33, to.Addr[12], to.Addr[13], to.Addr[14], to.Addr[15]}, ':'))
 		}
 		if from.Addr == nil {
 			from.Addr = net.IPv6unspecified
@@ -219,8 +228,7 @@ func (rc *RawConn) WriteTo(from, to *RawAddr, data []byte) (written int, err err
 	payload = append(payload, from.HardwareAddr...)
 	if rc.version == 4 {
 		// ETH ethertype
-		payload = append(payload, byte(syscall.ETH_P_IP>>8))
-		payload = append(payload, byte(syscall.ETH_P_IP&0xff))
+		payload = append(payload, byte(syscall.ETH_P_IP>>8), byte(syscall.ETH_P_IP&0xff))
 		// IP4 header
 		ilength, ulength := 28+len(data), 8+len(data)
 		payload = append(payload, []byte{
@@ -244,10 +252,10 @@ func (rc *RawConn) WriteTo(from, to *RawAddr, data []byte) (written int, err err
 			// UDP length + checksum (unused)
 			byte(ulength >> 8), byte(ulength), 0x00, 0x00,
 		}...)
+
 	} else {
 		// ETH ethertype
-		payload = append(payload, byte(syscall.ETH_P_IPV6>>8))
-		payload = append(payload, byte(syscall.ETH_P_IPV6&0xff))
+		payload = append(payload, byte(syscall.ETH_P_IPV6>>8), byte(syscall.ETH_P_IPV6&0xff))
 		// IP6 header
 		length := 8 + len(data)
 		payload = append(payload, []byte{
